Fix season_leaderboard view column type mismatch
sum() over a bigint column returns numeric in Postgres, which
conflicted with the existing bigint "points" column when replacing
the view. Cast back to bigint explicitly.

+++ supabase/migrations/20260623000000_perfect_card_bonus.sql
@@ -68,7 +68,7 @@ select
   extract(year from e.event_date)::int as season,
   el.user_id,
   el.nickname,
-  sum(el.points) as points
+  sum(el.points)::bigint as points
 from public.event_leaderboard el
 join public.events e on e.id = el.event_id
 group by season, el.user_id, el.nickname;
